tests/wpt/tests/css/css-text/parsing/word-break-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-break with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-word-break">
<meta name="assert" content="word-break supports only the grammar 'normal | break-all| keep-all | manual | auto-phrase | break-word'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("word-break", "normal");
test_valid_value("word-break", "break-all");
test_valid_value("word-break", "keep-all");
test_valid_value("word-break", "manual");
test_valid_value("word-break", "auto-phrase");
test_valid_value("word-break", "break-word");
</script>
</body>
</html>
